Improve strcutre of cassandra senssor testss
