+++ tests/providers/apache/cassandra/sensors/record.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+import unittest
+from unittest.mock import patch
+
+from airflow.providers.apache.cassandra.sensors.table import CassandraTableSensor
+
+
+class TestCassandraTableSensor(unittest.TestCase):
+    @patch("airflow.providers.apache.cassandra.sensors.table.CassandraHook")
+    def test_poke(self, mock_hook):
+        sensor = CassandraTableSensor(
+            task_id='test_task',
+            cassandra_conn_id='cassandra_default',
+            table='t',
+        )
+        sensor.poke(None)
+        mock_hook.return_value.table_exists.assert_called_once_with('t')
